Fix empty ordering in custom table
For some reason when the v-datatable is not ordered, vuetify prefills the sortDesc property with `[false]`. As a result the ordering filter is `undefined`, since the zip tries to zip the empty list from `sortBy` with `[false]`. So, first check if `sortBy` is empty to clear the ordering in that case.