Let the serverside table handle the sort options
It is very briddle to allow the component's user fiddling around with the options settings because it may become invalid. So now we are under the options propery control. This also allows us to define a converter for the ordering filer. So, whenever the sortDesc or sortBy changes in the options, the component will translate that to a DRF compatible format. It is also not necessary to check for empty sortBy's, since the zip automatically handles empty lists and thus not leaving the options settings invalid.